Move the conversion functions to their own module.
